+++ MundoBalloonApi.infrastructure/Data/Models/OrderProductsDetails.cs
@@ -5,8 +5,8 @@ public class OrderProductsDetails : BaseEntity
     public int OrderDetailsProductsId { get; init; }
     public int OrderId { get; init; }
     public int ProductVariantId { get; init; }
-    public decimal Quantity { get; init; }
-    public decimal Price { get; init; }
+    public double Quantity { get; init; }
+    public double Price { get; init; }
     public Orders? Order { get; set; }
     public ProductVariant? ProductVariant { get; set; }
 }
